refactor: extract components from large service files
Extracted reusable components from oversized service classes:

- ContentFilterService: Extracted SpamDetector and InappropriateWordFilter
- GameCharacterService: Extracted OfflineRewardCalculator and CharacterValidator
- ChatCacheService: Extracted RoomActivityTracker

Line count reductions:
- ContentFilterService: 508 -> 210 lines (-298)
- GameCharacterService: 515 -> 285 lines (-230)
- ChatCacheService: 500 -> 425 lines (-75)

New components created:
- SpamDetector: Spam detection logic (frequency, duplicates, caps, repetition, URL)
- InappropriateWordFilter: Inappropriate content filtering
- OfflineRewardCalculator: Offline reward calculation
- CharacterValidator: Character name validation
- RoomActivityTracker: Room activity tracking and analytics
